# backend/src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BLOG API",
    "description": "Description",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3300/"
    }
  ],
  "paths": {
...
    "/setting/addCategories": {
      "post": {
        "tags": [
          "Setting"
        ],
        "summary": "Add category",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/getAddCategoriesObject"
              },
              "example": {
                "categories": [
                  "카테고리1",
                  "카테고리2",
                  "카테고리3"
                ]
              }
            }
          }
        }
      }
    },
    "/setting/": {
      "post": {
        "tags": [
          "Setting"
        ],
        "summary": "Add Setting",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/getSettingCreateObject"
              },
              "example": {
                "type": "(...typeId)",
                "name": "category",
                "value": "new Category"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Setting"
        ],
        "summary": "put Setting",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/getSettingUpdateObject"
              },
              "example": {
                "_id": "(...settingId)",
                "type": "user",
                "role": "user",
                "userId": "userId",
                "name": "category",
                "value": "new Category"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Setting"
        ],
        "summary": "delete Setting",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/getSettingsDeleteObject"
              },
              "example": {
                "ids": [
                  "(...settingId)"
                ]
              }
            }
          }
        }
      }
    },
    "/setting/list": {
      "get": {
        "tags": [
          "Setting"
        ],
        "summary": "Get Setting",
        "description": "",
        "parameters": [
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ],
        "responses": {},
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Setting"
        ],
        "summary": "put Setting List",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/getSettingListUpdateObject"
              },
              "example": {
                "list": [
                  {
                    "_id": "(...settingId)",
                    "type": "user",
                    "role": "user",
                    "userId": "userId",
                    "name": "category",
                    "value": "new Category"
                  }
                ]
              }
            }
          }
        }
      }
    },
...
    "/visit/": {
      "get": {
        "tags": [
          "Visit"
        ],
        "summary": "Get visit",
        "description": "",
        "parameters": [
          {
            "$ref": "#/components/parameters/target"
          }
        ],
        "responses": {}
      },
      "post": {
        "tags": [
          "Visit"
        ],
        "summary": "post visit",
        "description": "",
        "parameters": [],
        "responses": {},
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/getVisitObject"
              },
              "example": {
                "target": "blog"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "getSettingUpdateObject": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "pattern": {
                    "type": "string",
                    "example": "^[a-fA-F0-9]{24}$"
                  },
                  "title": {
                    "type": "string",
                    "example": "설정 아이디"
                  }
                }
              },
              "type": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "pattern": {
                    "type": "string",
                    "example": "^[a-fA-F0-9]{24}$"
                  },
                  "title": {
                    "type": "string",
                    "example": "설정 타입"
                  }
                }
              },
              "role": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "pattern": {
                    "type": "string",
                    "example": "^[a-fA-F0-9]{24}$"
                  },
                  "title": {
                    "type": "string",
                    "example": "권한"
                  }
                }
              },
              "userId": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "pattern": {
                    "type": "string",
                    "example": "^[a-fA-F0-9]{24}$"
                  },
                  "title": {
                    "type": "string",
                    "example": "유저 아이디"
                  }
                }
              },
              "name": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "maxLength": {
                    "type": "number",
                    "example": 64
                  },
                  "minLength": {
                    "type": "number",
                    "example": 1
                  },
                  "title": {
                    "type": "string",
                    "example": "설정 이름"
                  }
                }
              },
              "value": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "maxLength": {
                    "type": "number",
                    "example": 64
                  },
                  "minLength": {
                    "type": "number",
                    "example": 1
                  },
                  "title": {
                    "type": "string",
                    "example": "설정 값"
                  }
                }
              }
            }
          },
          "required": {
            "type": "array",
            "example": [
              "_id",
              "name",
              "value"
            ],
            "items": {
              "type": "string"
            }
          },
          "additionalProperties": {
            "type": "boolean",
            "example": false
          },
          "title": {
            "type": "string",
            "example": "설정 갱신"
          }
        }
      },
      "getSettingListUpdateObject": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "list": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "array"
                  },
                  "items": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "object"
                      },
                      "properties": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "object",
                            "properties": {
                              "type": {
                                "type": "string",
                                "example": "string"
                              },
                              "pattern": {
                                "type": "string",
                                "example": "^[a-fA-F0-9]{24}$"
                              },
                              "title": {
                                "type": "string",
                                "example": "설정 아이디"
                              }
                            }
                          },
                          "type": {
                            "type": "object",
                            "properties": {
                              "type": {
                                "type": "string",
                                "example": "string"
                              },
                              "pattern": {
                                "type": "string",
                                "example": "^[a-fA-F0-9]{24}$"
                              },
                              "title": {
                                "type": "string",
                                "example": "설정 타입"
                              }
                            }
                          },
                          "role": {
                            "type": "object",
                            "properties": {
                              "type": {
                                "type": "string",
                                "example": "string"
                              },
                              "pattern": {
                                "type": "string",
                                "example": "^[a-fA-F0-9]{24}$"
                              },
                              "title": {
                                "type": "string",
                                "example": "권한"
                              }
                            }
                          },
                          "userId": {
                            "type": "object",
                            "properties": {
                              "type": {
                                "type": "string",
                                "example": "string"
                              },
                              "pattern": {
                                "type": "string",
                                "example": "^[a-fA-F0-9]{24}$"
                              },
                              "title": {
                                "type": "string",
                                "example": "유저 아이디"
                              }
                            }
                          },
                          "name": {
                            "type": "object",
                            "properties": {
                              "type": {
                                "type": "string",
                                "example": "string"
                              },
                              "maxLength": {
                                "type": "number",
                                "example": 64
                              },
                              "minLength": {
                                "type": "number",
                                "example": 1
                              },
                              "title": {
                                "type": "string",
                                "example": "설정 이름"
                              }
                            }
                          },
                          "value": {
                            "type": "object",
                            "properties": {
                              "type": {
                                "type": "string",
                                "example": "string"
                              },
                              "maxLength": {
                                "type": "number",
                                "example": 64
                              },
                              "minLength": {
                                "type": "number",
                                "example": 1
                              },
                              "title": {
                                "type": "string",
                                "example": "설정 값"
                              }
                            }
                          }
                        }
                      },
                      "required": {
                        "type": "array",
                        "example": [
                          "_id",
                          "name",
                          "value"
                        ],
                        "items": {
                          "type": "string"
                        }
                      },
                      "additionalProperties": {
                        "type": "boolean",
                        "example": false
                      },
                      "title": {
                        "type": "string",
                        "example": "설정 갱신"
                      }
                    }
                  },
                  "minItems": {
                    "type": "number",
                    "example": 1
                  },
                  "maxItems": {
                    "type": "number",
                    "example": 1000
                  },
                  "title": {
                    "type": "string",
                    "example": "설정 갱신 배열"
                  }
                }
              }
            }
          },
          "additionalProperties": {
            "type": "boolean",
            "example": false
          },
          "title": {
            "type": "string",
            "example": "설정 갱신 리스트"
          }
        }
      },
...
      "getAddCategoriesObject": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "categories": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "array"
                  },
                  "items": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      },
                      "minLength": {
                        "type": "number",
                        "example": 1
                      },
                      "maxLength": {
                        "type": "number",
                        "example": 40
                      },
                      "title": {
                        "type": "string",
                        "example": "카테고리"
                      }
                    }
                  },
                  "minItems": {
                    "type": "number",
                    "example": 1
                  },
                  "maxItems": {
                    "type": "number",
                    "example": 1000
                  },
                  "title": {
                    "type": "string",
                    "example": "카테고리 리스트"
                  }
                }
              }
            }
          },
          "required": {
            "type": "array",
            "example": [
              "categories"
            ],
            "items": {
              "type": "string"
            }
          },
          "additionalProperties": {
            "type": "boolean",
            "example": false
          },
          "title": {
            "type": "string",
            "example": "카테고리 리스트 추가"
          }
        }
      },
      "getSettingsDeleteObject": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "ids": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "array"
                  },
                  "items": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      },
                      "pattern": {
                        "type": "string",
                        "example": "^[a-fA-F0-9]{24}$"
                      },
                      "title": {
                        "type": "string",
                        "example": "설정 아이디"
                      }
                    }
                  },
                  "title": {
                    "type": "string",
                    "example": "설정 아이디 배열"
                  }
                }
              }
            }
          },
          "additionalProperties": {
            "type": "boolean",
            "example": false
          },
          "title": {
            "type": "string",
            "example": "설정 삭제"
          }
        }
      },
      "getSettingCreateObject": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "type": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "pattern": {
                    "type": "string",
                    "example": "^[a-fA-F0-9]{24}$"
                  },
                  "title": {
                    "type": "string",
                    "example": "설정 타입"
                  }
                }
              },
              "role": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "pattern": {
                    "type": "string",
                    "example": "^[a-fA-F0-9]{24}$"
                  },
                  "title": {
                    "type": "string",
                    "example": "권한"
                  }
                }
              },
              "userId": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "pattern": {
                    "type": "string",
                    "example": "^[a-fA-F0-9]{24}$"
                  },
                  "title": {
                    "type": "string",
                    "example": "유저 아이디"
                  }
                }
              },
              "name": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "maxLength": {
                    "type": "number",
                    "example": 64
                  },
                  "minLength": {
                    "type": "number",
                    "example": 1
                  },
                  "title": {
                    "type": "string",
                    "example": "설정 이름"
                  }
                }
              },
              "value": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "maxLength": {
                    "type": "number",
                    "example": 64
                  },
                  "minLength": {
                    "type": "number",
                    "example": 1
                  },
                  "title": {
                    "type": "string",
                    "example": "설정 값"
                  }
                }
              }
            }
          },
          "required": {
            "type": "array",
            "example": [
              "type",
              "name",
              "value"
            ],
            "items": {
              "type": "string"
            }
          },
          "additionalProperties": {
            "type": "boolean",
            "example": false
          },
          "title": {
            "type": "string",
            "example": "설정 추가"
          }
        }
      },
      "getVisitObject": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "target": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "minLength": {
                    "type": "number",
                    "example": 4
                  },
                  "maxLength": {
                    "type": "number",
                    "example": 24
                  },
                  "title": {
                    "type": "string",
                    "example": "방문 대상"
                  }
                }
              }
            }
          },
          "required": {
            "type": "array",
            "example": [
              "target"
            ],
            "items": {
              "type": "string"
            }
          },
          "additionalProperties": {
            "type": "boolean",
            "example": false
          },
          "title": {
            "type": "string",
            "example": "방문"
          }
        }
      },
...
    },
    "parameters": {
...
      "limit": {
        "in": "query",
        "name": "limit",
        "required": true,
        "description": "가져올 리스트 갯수",
        "schema": {
          "type": "number"
        }
      },
      "offset": {
        "in": "query",
        "name": "offset",
        "required": true,
        "description": "리스트 스킵 갯수",
        "schema": {
          "type": "number"
        }
      },
...
      "target": {
        "in": "query",
        "name": "target",
        "required": true,
        "description": "방문 대상",
        "schema": {
          "type": "string"
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}